+++ src/SeablastView.php
@@ -41,7 +41,7 @@ class SeablastView
             $this->renderJson($this->params->rest);
         } elseif (!isset($this->params->redirectionUrl)) {
             // HTML UI
-            if ($this->params->httpCode >= 400) {
+            if (isset($this->params->httpCode) && $this->params->httpCode >= 400) {
                 $this->model->mapping['template'] = 'error';
             }
             $this->renderLatte();