+++ CHANGELOG.md
@@ -21,6 +21,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### `Security` in case of vulnerabilities
 
+## [0.2.9] - 2025-03-09
+
+chore: PHP/8.4 support added
+
+### Added
+
+- support for PHP/8.4, including GitHub tests
+- support for PHP/8.0 added again
+- blast.sh - Management script for deployment and development of a Seablast application
+
+### Changed
+
+- mapping['roleIds'] exploded into array of integers is safer even thou in_array comparison is loose (so 2 equals " 2")
+
 ## [0.2.8] - 2025-02-09
 
 Strict versions of database adapter calls.
@@ -280,7 +294,8 @@ SeablastMysqli error logging improved, HTTPS identified
 - model returns knowledge()
 - a nice Under construction page
 
-[Unreleased]: https://github.com/WorkOfStan/seablast/compare/v0.2.8...HEAD?w=1
+[Unreleased]: https://github.com/WorkOfStan/seablast/compare/v0.2.9...HEAD?w=1
+[0.2.9]: https://github.com/WorkOfStan/seablast/compare/v0.2.8...v0.2.9?w=1
 [0.2.8]: https://github.com/WorkOfStan/seablast/compare/v0.2.7...v0.2.8?w=1
 [0.2.7]: https://github.com/WorkOfStan/seablast/compare/v0.2.6...v0.2.7?w=1
 [0.2.6]: https://github.com/WorkOfStan/seablast/compare/v0.2.5...v0.2.6?w=1